Fix backend issue
When sunpump reached a set amount of memes (maybe 100) the image class name is than being updated on the Sunpump website and this would result in the miniMEMEscraper not being able to scrape and prepare images from that location. the class name would than need to be updated manually in the popup.js file.

This fix makes sure that the miniMEME server continues to scrape and prepare memes even when the class name on Sunpump was changed.